offer cleard the float to decimal

+++ products/models.py
@@ -5,12 +5,12 @@ from category.models import Category
 class Products(models.Model):
     name = models.CharField(max_length=30)
     category = models.ForeignKey(Category,on_delete=models.CASCADE,related_name='products')
-    price = models.FloatField(default=0)
+    price = models.DecimalField(default=0, max_digits=20,decimal_places=2)
 
     def make_price(self):
         return 0.0
 
-    original_price = models.FloatField(default= make_price)
+    original_price = models.DecimalField(default= make_price, max_digits=20,decimal_places=2)
     stocks = models.IntegerField(default=0)
     discription = models.TextField(default='NONE')
